optimize

+++ src/app/controller/mouse/mouse_controller.cpp
@@ -1,7 +1,7 @@
 #include "mouse_controller.hpp"
 
 namespace NugieApp {
-  glm::vec3 MouseController::rotateInPlaceXZ(GLFWwindow* window, float dt, glm::vec3 currentCameraDirection, bool* isPressed) {
+  glm::vec3 MouseController::rotateInPlaceXZ(GLFWwindow* window, double dt, glm::vec3 currentCameraDirection, bool* isPressed) {
     glm::vec3 newCameraDirection = currentCameraDirection;
 
     if (glfwGetMouseButton(window, this->keymaps.rightButton) == GLFW_PRESS) {
@@ -14,7 +14,7 @@ namespace NugieApp {
       glfwGetCursorPos(window, &curDragged_x, &curDragged_y);
 
       if (!this->isFirstPressed) {
-        float theta = glm::radians((curDragged_y - this->lastDragged_y) * dt * this->lookSpeed * -1.0f);
+        float theta = glm::radians((curDragged_y - this->lastDragged_y) * dt * this->lookSpeed * -1.0);
         float phi = glm::radians((curDragged_x - this->lastDragged_x) * dt * this->lookSpeed);
 
         newCameraDirection.x = glm::cos(phi) * currentCameraDirection.x + glm::sin(phi) * currentCameraDirection.z;
